Refactor mainloop a bit more
